# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FlyRank Secure Auth API",
    "version": "1.0.0",
    "description": "A secure Node.js API with Supabase JWT Authentication."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "User created" }, "400": { "description": "Bad Request" } }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Authenticate user & return JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Login successful" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Terminate the user session",
        "security": [ { "bearerAuth": [] } ],
        "responses": { "204": { "description": "No Content" } }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Read public, unprotected data",
        "responses": { "200": { "description": "Success" } }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Read private user profile data",
        "security": [ { "bearerAuth": [] } ],
        "responses": { "200": { "description": "Success" }, "401": { "description": "Unauthorized" } }
      }
    }
  }
}